Fix the math symbols
Fix the math symbols so it is possible to use different signs in one expression, add or minus or times
or divide numbers in units, tens, hundreds, and so on. Also used indexes of results of expressions
to track the answers and use them when you are deleting some numbers from the expression.